refactor: split userService into separate files
