+++ packages/api/server/services/user/getPermissions.ts
@@ -1,0 +1,7 @@
+import permissionModel from '#server/models/permissionModel';
+import { Permissions } from '#server/models/permissionModel/types/Permissions.d';
+
+export default async (userId: number): Promise<Permissions> => {
+    const hash = await permissionModel.find([userId]);
+    return hash[userId] || {};
+};
