+++ packages/api/server/services/user/index.ts
@@ -1,9 +1,15 @@
+import create from './create';
 import deactivateInactiveUsers from './deactivateInactiveUsers';
+import getPermissions from './getPermissions';
+import register from './register';
 import sendInactiveUserAlerts from './sendInactiveUserAlerts';
 import updatePermissionOptions from './updatePermissionOptions';
 
 export default {
+    create,
     deactivateInactiveUsers,
+    getPermissions,
+    register,
     sendInactiveUserAlerts,
     updatePermissionOptions,
 };
